[refactor] 거리 타입 변환 및 허브 응답 dto수정

+++ order-service/src/main/java/com/shoonglogitics/orderservice/domain/delivery/application/dto/ListDeliveryRouteResult.java
@@ -14,9 +14,9 @@ public record ListDeliveryRouteResult(
 	UUID departureHubId,
 	UUID arrivalHubId,
 	Integer sequence,
-	Long estimatedDistance,
+	Integer estimatedDistance,
 	Integer estimatedDuration,
-	Long actualDistance,
+	Integer actualDistance,
 	Integer actualDuration,
 	DeliveryStatus status
 ) {
