+++ order-service/src/main/java/com/shoonglogitics/orderservice/domain/delivery/application/service/dto/RoutesInfo.java
@@ -6,14 +6,14 @@ public record RoutesInfo(
 	UUID departureHubId,
 	UUID arrivalHubId,
 	int sequence,
-	Long estimateDistance,
+	int estimateDistance,
 	int estimatedDuration
 ) {
 	public static RoutesInfo from(
 		UUID departureHubId,
 		UUID arrivalHubId,
 		int sequence,
-		Long estimateDistance,
+		int estimateDistance,
 		int estimatedDuration
 	) {
 		return new RoutesInfo(
