+++ order-service/src/main/java/com/shoonglogitics/orderservice/domain/delivery/presentation/dto/ProcessHubShippingCommand.java
@@ -8,7 +8,7 @@ public record ProcessHubShippingCommand(
 	UUID deliveryId,
 	UUID deliveryRouteId,
 	Boolean isDeparture,
-	Long distance,
+	Integer distance,
 	Integer duration,
 	AuthUser authUser
 ) {
@@ -16,7 +16,7 @@ public record ProcessHubShippingCommand(
 		UUID deliveryId,
 		UUID deliveryRouteId,
 		Boolean isDeparture,
-		Long distance,
+		Integer distance,
 		Integer duration,
 		AuthUser authUser
 	) {
